Added server to NodeMCU.
Changes:

Added a server side to the NodeMCU to allow remote access from a browser. Added "diagnostic" information on request. This only shows the current time, the alarm config, and the lights activity. Ideally, with this capability, the main server (where the alarm is obtained) should just tell the NodeMCU when to switch the lights on or off without any querries.