Cut down the HTML code, and took pressure of RAM.
Changes:

Basically, the HTML code being generated was not being nice to the RAM. Just to make things simpler and easier, the device is not controlled from the server, as it is easier to control all devices from one location. As such, there is no need for the HTML code to ask for forms/buttons/options. This just adds unnecessary complexity. Only the diagnostic data is displayed as HTML.

A buffer was added to store the remaining HTML code. This should prevent having to assign memory every time the HTML code is loaded/generated. I thought I was doing that with `static char*`, but I was wrong.

HTML code generator is now a `void` function, and the HTML buffer is `external`. This prevents passingvalues around, and means the HTML buffer is public.

All this has stabilised the RAM. Continuous reloading of the local device's HTML seems to stabilise the ram afte a few reloads. Continuous sending of light colour also stabilises after a few transmissions. A lot of ram is still being taken up between memory checks 5 and 9.