Removed all instances of `String` and replaced with a string manipulation library that uses char arrays.
Changes:

Added the "is alarm active" and "is alarm enabled" parameters to the diagnostics HTML code.

Created reusable char arryas containing the API request and parameter names to search for when requesting from device. This reduced the stress on RAM (or attempts to).

Added a string manipulation library to avoid having to play with `String` objects. This library uses char arrays and low-level array manipulation instead. This has significantly helped RAM usage.